+++ app/views/userManagement/verificationTeam.php
@@ -3,7 +3,8 @@ $header = new HTMLHeader("Verification Team");
 $nav = new Navigation('management');
 $sidebar = new SidebarNav("user", "verification");
 $basePage = BASEURL . '/userManagement/verificationTeam';
-$base = BASEURL . '/userManagement' ;
+$base = BASEURL . '/userManagement';
+
 
 ?>
 <main class=" full-width ">
@@ -12,23 +13,32 @@ $base = BASEURL . '/userManagement' ;
             <div class="row no-gap">
                 <div class="col-12 left fill-container">
                     <h1 class="header-1 ">
-                        Verification Team Management
+                        Verification Team
                     </h1>
                 </div>
 
                 <div class="col-9 col-small-10 col-large-9 fill-container left">
-                    <?php 
-                     if (isset($data['result'])) {
+
+                    <?php
+                    if (isset($data['result'])) {
                         new SearchUser("Verification Team");
-                    }?>
+                    }
+
+
+                    ?>
                 </div>
                 <div class="col-1 display-small-none"></div>
                 <div class="col-2 col-large-3 fill-container right">
-                    <button class="bg-blue white border-rounded header-nb padding-3 right"
-                        onclick=" location.href='<?= $base ?>/create/verificationTeam'">
-                        <i data-feather="user-plus" class=" vertical-align-middle "></i>
-                        <span class="display-large-inline-block padding-left-2 display-none">Add User</span>
-                    </button>
+                    <?php
+                    if ($_SESSION['role'] == 'Admin') {
+                        echo "
+                <button class='bg-blue-hover white border-rounded header-nb padding-3 right'
+                        onclick='location.href=\"$base/create/verificationTeam\"'>
+                        <i data-feather='user-plus' class=' vertical-align-middle '></i>
+                        <span class='display-large-inline-block padding-left-2 display-none'>Add User</span>
+                    </button>";
+                    }
+                    ?>
                 </div>
             </div>
 
@@ -39,11 +49,11 @@ $base = BASEURL . '/userManagement' ;
                 <div class=" shadow-small border-rounded-more   fill-container col-12 ">
                     <div class="row no-gap fill-container">
                         <div id="table"
-                            class="col-8 col-small-9 table fill-container border-rounded-more-left padding-top-4 padding-bottom-5 shadow-small bg-white ">
-                            <div class="hs padding-horizontal-5 padding-vertical-3"
+                            class="col-8 col-small-9 table fill-container border-rounded-more-left padding-top-4 padding-bottom-5 shadow-small bg-white "
                             data-current-page="<?php if (isset($data['page'])) {
                                 echo $data['page'];
                             } ?>" aria-live="polite">
+                            <div class="hs padding-horizontal-5 padding-vertical-3">
                                 <div class="col-2  text-overflow bold">First Name</div>
                                 <div class="col-2  text-overflow bold">Last Name</div>
                                 <div class="col-2  text-overflow bold">User Name</div>
@@ -56,16 +66,22 @@ $base = BASEURL . '/userManagement' ;
                             </div>
                             <?php
                             if (isset($data['result'])) {
-                                $result = array_slice($data['result'], ($data['page'] - 1) * $data['perPage'], $data['perPage']);
 
                                 $useridArray = array();
-                                foreach ($result as $key => $value) {
+                                foreach ($data['result'] as $key => $value) {
                                     array_push($useridArray, $value->UserId);
                                     $gender = $value->Gender;
                                     if ($gender == 'm') {
                                         $gender = 'Male';
                                     } else {
                                         $gender = 'Female';
+                                    }
+                                    $verifiedStatus = $value->VerifiedStatus;
+                                    if ($verifiedStatus == "verified") {
+                                        $verifiedstyle = " border-accent accent";
+                                    } else {
+                                        $verifiedstyle = " border-grey grey";
+                                        $verifiedStatus = "not verified";
                                     }
                                     echo "<div class='hs list-items padding-horizontal-5 padding-vertical-2 border-1 border-white'>";
                                     echo "<div class='person-name col-2  text-overflow ' title='" . $value->FirstName . "' >" . $value->FirstName . "</div>";
@@ -89,21 +105,35 @@ $base = BASEURL . '/userManagement' ;
 
                             <?php
                             if (isset($useridArray)) {
-                                foreach ($useridArray as $userid) {
-                                    echo "<div class='row less-gap padding-1 padding-horizontal-3 list-item-action'>";
-                                    echo "<div class='col-6 fill-container '>";
-                                    echo "<a href='" . $base . "/userEdit/verificationTeam/$userid'><div class=' fill-container border-blue bg-white blue-hover border-1 border-rounded padding-vertical-1  center'>";
-                                    echo "<i data-feather='edit' class='feather-body display-inline-block display-small-none'></i> <span class='display-small-block  display-none'>Edit</span>";
-                                    echo "</div></a>";
-                                    echo "</div>";
+                                if ($_SESSION['role'] == 'Admin') {
 
-                                    echo "<div class='col-6 fill-container '>";
-                                    echo "<a onclick='deleteUser($userid)' class='cursor-pointer'><div class=' fill-container border-red bg-white red-hover border-1 border-rounded padding-vertical-1  center'>";
-                                    echo "<i data-feather='trash' class='feather-body display-inline-block display-small-none'></i> <span class='display-small-block  display-none'>Delete</span>";
-                                    echo "</div></a>";
+                                    foreach ($useridArray as $userid) {
+                                        echo "<div class='row less-gap padding-1 padding-horizontal-3 list-item-action'>";
+                                        echo "<div class='col-6 fill-container '>";
+                                        echo "<a href='" . $base . "/userEdit/verificationTeam/$userid'><div class=' fill-container border-blue bg-white blue-hover border-1 border-rounded padding-vertical-1  center'>";
+                                        echo "<i data-feather='edit' class='feather-body display-inline-block display-small-none'></i> <span class='display-small-block  display-none'>Edit</span>";
+                                        echo "</div></a>";
+                                        echo "</div>";
 
-                                    echo "</div>";
-                                    echo "</div>";
+                                        echo "<div class='col-6 fill-container '>";
+                                        echo "<a onclick='deleteUser($userid)' class='cursor-pointer'><div class=' fill-container border-red bg-white red-hover border-1 border-rounded padding-vertical-1  center'>";
+                                        echo "<i data-feather='trash' class='feather-body display-inline-block display-small-none'></i> <span class='display-small-block  display-none'>Delete</span>";
+                                        echo "</div></a>";
+
+                                        echo "</div>";
+                                        echo "</div>";
+                                    }
+                                }else{
+                                    foreach ($useridArray as $userid) {
+                                        echo "<div class='row less-gap padding-1 padding-horizontal-3 list-item-action'>
+                                                <div class='col-12 fill-container '>
+                                                    <div class=' cursor-default fill-container border-blue bg-white blue-hover border-1 border-rounded padding-vertical-1  center'>
+                                                        <i data-feather='alert-circle' class='feather-body display-inline-block display-small-none'></i> 
+                                                        <span class='display-small-block  display-none'>No Actions</span>
+                                                    </div>
+                                                </div>
+                                             </div>"; 
+                                    }
                                 }
                             }
                             ?>
@@ -162,7 +192,7 @@ $base = BASEURL . '/userManagement' ;
                                             } else {
                                                 echo '<option value="100">100</option>';
                                             }
-                                        }else{
+                                        } else {
                                             echo '<option value="' . $data['perPage'] . '" selected>' . $data['perPage'] . '</option>';
                                             echo '<option value="2">2</option>';
                                             echo '<option value="5">5</option>';
@@ -172,7 +202,7 @@ $base = BASEURL . '/userManagement' ;
                                             echo '<option value="80">80</option>';
                                             echo '<option value="100">100</option>';
                                         }
- 
+
                                     }
                                     ?>
                                 </select>
@@ -198,7 +228,6 @@ $base = BASEURL . '/userManagement' ;
         </div>
     </div>
 </main>
-
 
 <script>
     function deleteUser(id) {
